backend/requirements.txt
fastapi
uvicorn[standard]
aiohttp
python-dotenv
sqlalchemy
psycopg2-binary
